changed some doubles to long where appropriate

+++ src/main/scala/elaio/neuralnet/processing/NeuronCollectionCache.scala
@@ -4,12 +4,12 @@ import scala.collection.mutable.HashMap
 import elaio.neuralnet.units.Neuron
 
 object NeuronCollectionCache {
-  private val cache: HashMap[Double, Neuron] = HashMap.empty[Double, Neuron]
+  private val cache: HashMap[Long, Neuron] = HashMap.empty[Long, Neuron]
 
   def clear(): Unit = cache.clear()
 
   def add(neuron: Neuron): Unit = cache(neuron.id) = neuron
 
-  def get(id: Double): Neuron = cache.getOrElse(id, null)
+  def get(id: Long): Neuron = cache.getOrElse(id, null)
 
 }
